LangchainBackend/requirements.txt
langchain==0.1.12
langchain-community>=0.0.28,<0.1
langchain-core>=0.1.31,<0.2.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
httpx==0.27.0
pymongo==4.6.1
langchain-mongodb==0.1.1 